Enable Single Sign on via Portal session
Adds support for a new websso endpoint that will be used to log a user
in via the existing Portal session. Since the Keystone credentials are
equivalent to the Portal ones, we perform a shadow login on behalf of
the user from Portal. Redirect logged out users to Portal to facilitate
the automatic logging in based on their Portal session.

Ticket: https://collab.tacc.utexas.edu/issues/15088