Detect region when using versionless id endpoints
When using versionless Keystone endpoints in the service catalog (which
is the recommended way as it allows the client to do version selection),
region detection was broken. The reason is that a v3 suffix was
automatically added to the identity service URL before doing the region
detection. Region detection works by querying the service catalog for
which endpoint matches the Keystone auth URL used for the request. Since
the endpoint URL does not have a version suffix, the search will never
succeed, so the region defaults to None, and then later the first
endpoint is picked as a fallback behavior.

This fixes the check to also look for the original, unfixed URL when
trying to detect the region.