Rewrite custom SSO as edge case of Keystone SSO
It turns out that, upon close inspection, our custom SSO flow is
remarkably similar to the "default WebSSO redirect" flow added in the
Stein release of Horizon. We just need to add support for one bit:
changing the redirect mechanism to allow an operator to configure an
absolute URL, as opposed to one created by a Keystone/IdP protocol pair.
The rest of the flow is the same: the user is redirected to a page via a
GET request (with the inbound origin appended as a param), and then the
/websso endpoint in Horizon takes a POST request from the IdP with the
Keystone token of the user, which is then used to initialize the user's
session.

This also adds support for a way to specify, even in the IdP chooser,
that you want to use the default redirect behavior--this is accomplished
by using "None" as the IdP in the WEBSSO_IDP_MAPPING, like so:

WEBSS_IDP_MAPPING = {
  'default': (None, None),
  'some_idp': ('idp', 'saml'),
}