Change Portal SSO to special case of federation (#15)
* Rewrite custom SSO as edge case of Keystone SSO

It turns out that, upon close inspection, our custom SSO flow is
remarkably similar to the "default WebSSO redirect" flow added in the
Stein release of Horizon. We just need to add support for one bit:
changing the redirect mechanism to allow an operator to configure an
absolute URL, as opposed to one created by a Keystone/IdP protocol pair.
The rest of the flow is the same: the user is redirected to a page via a
GET request (with the inbound origin appended as a param), and then the
/websso endpoint in Horizon takes a POST request from the IdP with the
Keystone token of the user, which is then used to initialize the user's
session.

This also adds support for a way to specify, even in the IdP chooser,
that you want to use the default redirect behavior--this is accomplished
by using "None" as the IdP in the WEBSSO_IDP_MAPPING, like so:

WEBSS_IDP_MAPPING = {
  'default': (None, None),
  'some_idp': ('idp', 'saml'),
}

Instead, go back to the login page, since we didn't actually use the
redirect flow.

* Support redirect 'next' param in WebSSO

If a user logs in via WebSSO, the 'next' parameter is not respected, and
so the user ultimately must re-navigate to their target destination once
they complete their login. This fixes that by including the 'next'
parameter on the Keystone WebSSO start request. Keystone will ignore
this flag. However, we can then read it from the referer when Horizon is
handling its final part of the login flow.

* Add page to display if no default logout URL

If there is a default redirect URL, but no default logout URL, the user
can effectively not logout, as they will be brought back to the login
page, which will automatically log them in again. This adds a logout
page with some overriddable messaging, and a link to do a "deep" logout

* Support passing redirect on websso as query param

Instead of extracting from the referer, sometimes the triggering service
may choose to include this directly.