Generate v3oidc openrc files for federated users
When the user is logging in via federation, it no longer makes sense to
generate v3password-style RC files, as they will probably not be able to
log in with this method. Instead, generate a placeholder configuration
that works with some supported method, in this default, v3oidcpassword.

This should probably be configurable long-term somehow.