Redirect federated users to special logout URL
This slightly changes how the legacy Portal SSO implementation works...
the logout page now shows a URL configurable via the

  WEBSSO_DEFAULT_REDIRECT_LOGOUT_CONFIRM_URL

setting, which is a URL string. This URL is only used when

  WEBSSO_DEFAULT_REDIRECT_LOGOUT_CONFIRM

is set to True.

This also updates the logout condition to check if the user should be
getting the new logout flow. Any federated user should be getting the
new logout flow; it does not make sense to do it otherwise.